requirements.txt
Django==4.2.13
djangorestframework==3.15.1
python-telegram-bot==13.7
python-dotenv==1.0.0
requests